src/requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.4
dataclasses==0.6
Flask==2.0.3
gunicorn==20.1.0
html-table-parser-python3==0.2.0
idna==3.3
importlib-metadata==4.8.3
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
pandas==1.1.5
PyMySQL==1.0.2
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
requests==2.27.1
six==1.16.0
typing_extensions==4.1.1
urllib3==1.26.9
Werkzeug==2.0.3
zipp==3.6.0
